+++ resources/views/livewire/distributer-lokacija.blade.php
@@ -37,11 +37,7 @@
                                         <td class="px-2 py-2">{{ $item->adresa }}</td>
                                         <td class="px-2 py-2">{{ $item->mesto }}</td>  
                                         <td class="px-2 py-2">{{ $item->r_naziv }}</td>                                       
-                                        <td class="px-2 py-2 text-right">
-                                            <x-jet-danger-button class="mx-2" wire:click="deleteShowModal({{ $item->id }}, '{{ $item->l_naziv }}')" title="Obriši lokaciju">
-                                                <svg class="fill-current w-4 h-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M135.2 17.69C140.6 6.848 151.7 0 163.8 0H284.2C296.3 0 307.4 6.848 312.8 17.69L320 32H416C433.7 32 448 46.33 448 64C448 81.67 433.7 96 416 96H32C14.33 96 0 81.67 0 64C0 46.33 14.33 32 32 32H128L135.2 17.69zM31.1 128H416V448C416 483.3 387.3 512 352 512H95.1C60.65 512 31.1 483.3 31.1 448V128zM111.1 208V432C111.1 440.8 119.2 448 127.1 448C136.8 448 143.1 440.8 143.1 432V208C143.1 199.2 136.8 192 127.1 192C119.2 192 111.1 199.2 111.1 208zM207.1 208V432C207.1 440.8 215.2 448 223.1 448C232.8 448 240 440.8 240 432V208C240 199.2 232.8 192 223.1 192C215.2 192 207.1 199.2 207.1 208zM304 208V432C304 440.8 311.2 448 320 448C328.8 448 336 440.8 336 432V208C336 199.2 328.8 192 320 192C311.2 192 304 199.2 304 208z"/></svg>
-                                            </x-jet-button>
-                                        </td>
+                                        <td class="px-2 py-2 text-right"></td>
                                     </tr>
                                 @endforeach
                             @else 
@@ -61,42 +57,6 @@
     </div>
 
 
-
-    {{-- The Delete Modal --}}
-    <x-jet-dialog-modal wire:model.live="modalDeleteLocVisible">
-        <x-slot name="title">
-            {{ __('Uklanjanje lokacije distributera') }}
-        </x-slot>
-
-        <x-slot name="content">
-            <div class="py-4">Lokacija: <span class="font-bold">{{ $l_naziv }}</span></div>
-            @if($delete_error)
-                <div class="bg-red-50 border border-red-500 text-red-500 px-4 py-3 rounded relative my-4 " role="alert">
-                    <p class="">Greška!<br /></p>
-                    @foreach($delete_error as $errtxt )
-                    <span class="font-bold block sm:inline">{{ $errtxt }}</span><br />
-                    @endforeach
-                    <span class="absolute top-0 bottom-0 right-0 px-4 py-3">
-                        <svg class="fill-red-500 h-6 w-6 " xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M506.3 417l-213.3-364c-16.33-28-57.54-28-73.98 0l-213.2 364C-10.59 444.9 9.849 480 42.74 480h426.6C502.1 480 522.6 445 506.3 417zM232 168c0-13.25 10.75-24 24-24S280 154.8 280 168v128c0 13.25-10.75 24-23.1 24S232 309.3 232 296V168zM256 416c-17.36 0-31.44-14.08-31.44-31.44c0-17.36 14.07-31.44 31.44-31.44s31.44 14.08 31.44 31.44C287.4 401.9 273.4 416 256 416z"/></svg>
-                    </span>
-                    
-                </div>
-            @else
-                <div class="font-bold text-red-500"> Da li ste sigurni da želite da uklonite ovu lokaciju?</div>
-            @endif
-        </x-slot>
-
-        <x-slot name="footer">
-            <x-jet-secondary-button wire:click="$toggle('modalDeleteLocVisible')" wire:loading.attr="disabled">
-                {{ __('Otkaži') }}
-            </x-jet-secondary-button>
-            @if(!$delete_error)
-                <x-jet-danger-button class="ml-2" wire:click="delete" wire:loading.attr="disabled">
-                    {{ __('Ukloni lokaciju') }}
-                </x-jet-danger-button>
-            @endif
-        </x-slot>
-    </x-jet-dialog-modal>
 
     {{--  MODAL DOAJ LOKACIJU --}}
     <x-jet-dialog-modal wire:model.live="modalAddLocVisible">